# swagger.yaml
openapi: 3.0.3
info:
  title: Elice SW Track 7th Web Project Team 6
  description: 도서 판매 사이트 API 문서
  version: 1.0.0
servers:
  - url: http://localhost:5000/api/v1
...
paths:
...
  /users/{userId}:
    get:
      tags:
        - users
      summary: 회원 정보
      description: 회원정보를 가져오는 API
      security:
        - BearerAuth: []
      parameters:
        - name: userId
          in: path
          description: 유저 ID
          required: true
          schema:
            type: string
      responses:
        "200":
          description: 성공**(name, address가 없을경우 name, zipCode, detail은 ""로 반환)**
          content:
            application/json:
              schema:
                type: object
                properties:
                  _id:
                    type: string
                    example: asdjvasdvnecnsjdfl
                  email:
                    type: string
                    example: example@example.com
                  phoneNumber:
                    type: string
                    example: 010-1234-1234
                  authority:
                    type: string
                    example: 일반
                  name:
                    type: string
                    example: 오강산
                  address:
                    type: object
                    properties:
                      zipCode:
                        type: string
                        example: 12345
                      detail1:
                        type: string
                        example: 서울시 강남구 역삼동 123
                      detail2:
                        type: string
                        example: 403호
        "500":
          description: 서버 에러

    put:
      tags:
        - users
      summary: 회원 정보 수정
      description: 회원 정보를 수정하는 API
      security:
        - BearerAuth: []
      parameters:
        - name: userId
          in: path
          description: 유저 ID
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                password:
                  type: string
                  example: password
                phoneNumber:
                  type: string
                  example: 010-1234-1234
                name:
                  type: string
                  example: 오강산
                address:
                  type: object
                  properties:
                    zipCode:
                      type: string
                      example: 12345
                    detail1:
                      type: string
                      example: 서울시 강남구 역삼동 123
                    detail2:
                      type: string
                      example: 111호
        required: true
      responses:
        "200":
          description: 성공
        "400":
          description: 잘못된 요청
        "500":
          description: 서버 에러

    delete:
      tags:
        - users
      summary: 회원 정보 삭제
      description: 회원 탈퇴 API
      security:
        - BearerAuth: []
      parameters:
        - name: userId
          in: path
          description: 유저 ID
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                password:
                  type: string
                  example: password
      responses:
        "200":
          description: 성공
        "400":
          description: 잘못된 요청
        "500":
          description: 서버 에러
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer